Speed up `--update` option test
The variable `mock_update` is already defined with a pytest fixture. There is no need to create it again inside a `with`
bloc.

Furthermore, patching `sys.exit` ensures that the pytest command does not exit while running the test, but replacing
`sys.exit` with a no-op means that the program won't stop after the update. As a consequence, the test execute all the
attack modules. Using `pytest.raises` instead to catch the exception ensure the program ends after the update. This
reduces the execution time of the test from 3s to 0.6s.